[front] Move SharePoint folder title formatting to Microsoft provider
Keep provider-specific display logic out of generic content node helpers and
apply the SharePoint site name prefix in the agent builder knowledge views.
