+++ front/components/ContentNodeTree.tsx
@@ -12,10 +12,7 @@ import type { ReactNode } from "react";
 import React, { useCallback, useContext, useState } from "react";
 
 import { useSendNotification } from "@app/hooks/useNotification";
-import {
-  getDisplayTitleForContentNode,
-  getVisualForContentNode,
-} from "@app/lib/content_nodes";
+import { getVisualForContentNode } from "@app/lib/content_nodes";
 import { classNames } from "@app/lib/utils";
 import type { APIError, ContentNode } from "@app/types";
 
@@ -151,16 +148,9 @@ function ContentNodeTreeChildren({
     isLoadingMore,
   } = useResourcesHook(parentId);
 
-  const getNodeTitle = useCallback(
-    (node: ContentNode) => getDisplayTitleForContentNode(node),
-    []
-  );
-
   const filteredNodes = resources
-    .filter(
-      (n) => filter.trim().length === 0 || getNodeTitle(n).includes(filter)
-    )
-    .sort((a, b) => getNodeTitle(a).localeCompare(getNodeTitle(b)));
+    .filter((n) => filter.trim().length === 0 || n.title.includes(filter))
+    .sort((a, b) => a.title.localeCompare(b.title));
 
   const getCheckedState = useCallback(
     (node: ContentNode) => {
@@ -216,7 +206,6 @@ function ContentNodeTreeChildren({
 
       {filteredNodes.map((n) => {
         const checkedState = getCheckedState(n);
-        const title = getNodeTitle(n);
         return (
           <Tree.Item
             key={n.internalId}
@@ -224,7 +213,7 @@ function ContentNodeTreeChildren({
             type={
               showExpand === false ? "item" : n.expandable ? "node" : "leaf"
             }
-            label={title}
+            label={n.title}
             labelClassName={
               n.providerVisibility === "private"
                 ? "after:content-['(private)'] after:text-warning after:ml-1"
