+++ front/components/assistant/details/tabs/AgentInfoTab/AssistantKnowledgeSection.tsx
@@ -27,10 +27,7 @@ import type {
 } from "@app/lib/api/assistant/configuration/types";
 import { getContentNodeInternalIdFromTableId } from "@app/lib/api/content_nodes";
 import { getConnectorProviderLogoWithFallback } from "@app/lib/connector_providers_ui";
-import {
-  getDisplayTitleForDataSourceViewContentNode,
-  getVisualForDataSourceViewContentNode,
-} from "@app/lib/content_nodes";
+import { getVisualForDataSourceViewContentNode } from "@app/lib/content_nodes";
 import {
   canBeExpanded,
   getDisplayNameForDataSource,
@@ -547,7 +544,7 @@ function DataSourceViewSelectedNodes({
   return nodes.map((node) => (
     <Tree.Item
       key={node.internalId}
-      label={getDisplayTitleForDataSourceViewContentNode(node)}
+      label={node.title}
       type={node.expandable && viewType !== "table" ? "node" : "leaf"}
       visual={getVisualForDataSourceViewContentNode(node)}
       className="whitespace-nowrap"
