+++ front/lib/content_nodes.ts
@@ -156,74 +156,3 @@ export function getLocationForDataSourceViewContentNodeWithSpace(
     ? `${spaceName} › ${locationWithoutSpace}`
     : locationWithoutSpace;
 }
-
-const SHAREPOINT_SITE_NAME_REGEX =
-  /^https?:\/\/[^/]*sharepoint\.com\/(?::f:\/r\/)?(?:sites|teams)\/([^/?#]+)/i;
-
-function extractSharePointSiteNameFromSourceUrl(
-  sourceUrl: string
-): string | null {
-  const match = sourceUrl.match(SHAREPOINT_SITE_NAME_REGEX);
-  const encodedSiteName = match?.[1];
-  if (!encodedSiteName) {
-    return null;
-  }
-
-  try {
-    return decodeURIComponent(encodedSiteName);
-  } catch {
-    // decodeURIComponent throws URIError for malformed escape sequences.
-    return encodedSiteName;
-  }
-}
-
-/**
- * Display title used in knowledge selection trees.
- * For Microsoft SharePoint root folders, include the site name to disambiguate
- * repeated folder names across sites.
- */
-export function getDisplayTitleForDataSourceViewContentNode(
-  node: DataSourceViewContentNode
-): string {
-  if (node.dataSourceView.dataSource.connectorProvider !== "microsoft") {
-    return node.title;
-  }
-
-  if (
-    node.type !== "folder" ||
-    node.parentInternalId !== null ||
-    !node.sourceUrl
-  ) {
-    return node.title;
-  }
-
-  const siteName = extractSharePointSiteNameFromSourceUrl(node.sourceUrl);
-  return siteName ? `${siteName} → ${node.title}` : node.title;
-}
-
-export function getDisplayTitleForContentNode(node: ContentNode): string {
-  // ContentNodeTree sometimes receives DataSourceViewContentNode items (which carry
-  // connector metadata). When available, use the richer formatting.
-  return isDataSourceViewContentNode(node)
-    ? getDisplayTitleForDataSourceViewContentNode(node)
-    : node.title;
-}
-
-function isDataSourceViewContentNode(
-  node: ContentNode
-): node is DataSourceViewContentNode {
-  const candidate = node as unknown as { dataSourceView?: unknown };
-  if (
-    !candidate.dataSourceView ||
-    typeof candidate.dataSourceView !== "object"
-  ) {
-    return false;
-  }
-
-  const dataSourceView = candidate.dataSourceView as {
-    dataSource?: unknown;
-  };
-  return (
-    !!dataSourceView.dataSource && typeof dataSourceView.dataSource === "object"
-  );
-}
